+++ src/core/index.ts
@@ -3,3 +3,4 @@ export { createUnistylesImageBackground } from './createUnistylesImageBackground
 export { withUnistyles } from './withUnistyles'
 export { passForwardedRef } from './passForwardRef'
 export { getClassName } from './getClassname'
+export { useUnistyles } from './useUnistyles'
