+++ src/core/useProxifiedUnistyles/index.ts
@@ -1,0 +1,1 @@
+export { useProxifiedUnistyles } from './useProxifiedUnistyles'
