+++ src/core/useProxifiedUnistyles/types.ts
@@ -1,0 +1,7 @@
+import type { UnistyleDependency } from '../../specs'
+
+export type ListenerProps = {
+    updateTheme: VoidFunction,
+    updateRuntime: VoidFunction,
+    dependencies: Array<UnistyleDependency>
+}
