+++ src/core/useUnistyles.ts
@@ -1,0 +1,10 @@
+import { useProxifiedUnistyles } from './useProxifiedUnistyles'
+
+export const useUnistyles = () => {
+    const { proxifiedRuntime, proxifiedTheme } = useProxifiedUnistyles()
+
+    return {
+        theme: proxifiedTheme,
+        rt: proxifiedRuntime
+    }
+}
