+++ src/index.ts
@@ -1,6 +1,6 @@
 export { StyleSheet, UnistylesRuntime, StatusBar, NavigationBar } from './specs'
 export { mq } from './mq'
 export type { UnistylesThemes, UnistylesBreakpoints } from './global'
-export { withUnistyles } from './core'
+export { withUnistyles, useUnistyles } from './core'
 export type { UnistylesVariants } from './types'
 export { Display, Hide, ScopedTheme } from './components'
